mode Node class to new java class file
