+++ src/com/yandex/kanbanboard/service/InMemoryHistoryManager.java
@@ -75,39 +75,5 @@ public class InMemoryHistoryManager implements HistoryManager {
     }
 }
 
-class Node<Task> {
-    private Task value;
-    private Node<Task> prev;
-    private Node<Task> next;
 
-    public Node(Task value, Node<Task> prev, Node<Task> next) {
-        this.value = value;
-        this.prev = prev;
-        this.next = next;
-    }
-
-    public Task getValue() {
-        return value;
-    }
-
-    public Node<Task> getPrev() {
-        return prev;
-    }
-
-    public Node<Task> getNext() {
-        return next;
-    }
-
-    public void setNext(Node<Task> next) {
-        this.next = next;
-    }
-
-    public void setValue(Task value) {
-        this.value = value;
-    }
-
-    public void setPrev(Node<Task> prev) {
-        this.prev = prev;
-    }
-}
